added some irq flag reg experimenting (esp. timer)
this commit is quite senseless and for experimenting purposes only,
however it shows that the broadway processor interface irq #14
(regarding the flag register BW_PI_IRQFLAG) seems to be hard-wired
to the hollywood pic: whenever an interrupt occurs that is masked
in the HW_PPCIRQMASK register, both the corresponding flag in
HW_PPCIRQFLAG and bit 14 of BW_PI_IRQFLAG are set - for clearing
that processor interface bit again, the only thing you need to do
is actually clearing the flag in HW_PPCIRQFLAG, as the following
output shows (compiled with #define TEST_WITH_TIMER):

[before_flag_clear] ppcirqflag: 81010021
[before_flag_clear] pi irqflag: 00014100
 [after_flag_clear] ppcirqflag: 81010020
 [after_flag_clear] pi irqflag: 00010100

magically, after clearing the flag reg in the hollywood pic,
that processor interface bit 14 (0x000040000) is vanished.

however, if you try to do the same when an ohci interrupt occurs
(comment out the #define TEST_WITH_TIMER), clearing the corresponding
irq flag in the hollywood pic doesn't even work! output:

[before_flag_clear] ppcirqflag: 01000020
[before_flag_clear] pi irqflag: 00014100
[after_flag_clear] ppcirqflag: 01000020
[after_flag_clear] pi irqflag: 00014100

(expected ppcirqflag value is 01000000)
what that means? i don't know. i think because of the wrong
initialization of the ohci controller it never stops throwing
interrupts when it started once.

the timer interrupt occurs again because it is used for gecko
polling in mini, so there's nothing wrong here.